// src/types/EmployeeResponse.tsx
export interface EmployeeResponse {
  employeeId: number;
  employeeCode: string;
  firstName: string;
  lastName: string;
  gender: string;
  birthDate: string;
  hireDate: string;
  email: string;
  phoneNumber: string;
  departmentName: string;
  positionName: string;
  salary: number;
  status: string;
  profilePictureUrl: string;
  presignedRequestUrl: string;
}